generalize to openPollAndCallback
